created basic folder structure for the project
