+++ src/pages/EquipmentRental.tsx
